管理员端操作日志

+++ heima-leadnews-model/src/main/java/com/heima/model/admin/pojos/AdUserOperation.java
@@ -52,7 +52,7 @@ public class AdUserOperation implements Serializable {
      * 操作类型
      */
     @TableField("type")
-    private Short type;
+    private Integer type;
 
     /**
      * 操作描述
